New method.
Method for finding lines is probing the frame with pre-defined lines and cummulating the pixel values in one line. The line with the highest cumulated sum is the needle axes. The needle tip is found by moving an averaging window over the line and finding the maximal difference along the line.

function.py holds custom functions
parameters.py hold processing and other parameters
process_multiple_frames.py processes five frames and shows the steps involved
provess_multiple_frames_no_steps.py is equal to the above but does not show the processing steps